Fixed a pair of Codacy issues
Criticism and suggestions are always welcome!
