+++ src/instr/llvm_cfg_gen.py
@@ -37,7 +37,7 @@ def extract_function_identifiers(text):
     fnc_ident_pattern = '([A-Z][0-9A-Za-z]*)([ \t\n\r\f\v]*\()'
     res = set()
     for match in re.finditer(fnc_ident_pattern, text):
-        res.add(match.group(1))     
+        res.add(match.group(1))
     return res
 
 copyright_str = '// file generated by a ROOT-Sim build script, all rights reserved\n'
@@ -48,7 +48,7 @@ if __name__ == '__main__':
     with open(source_output, 'w') as f:
         f.write(copyright_str) # TODO do we need this?
 
-        f.write('#include <instr/instr_cfg.hpp>\n\n')
+        f.write('#include <instr/instr_llvm.hpp>\n\n')
 
         f.write('const char *const to_substitute[] = ')
         f.write(collection_to_c_array(instr_specs["to_substitute"], 1))
